backend/requirements.txt
uvicorn
fastapi
PyJWT
sqlalchemy
psycopg2-binary
httpx
google-auth
dotenv
scikit-learn
requests
tiktoken
psutil
pyupbit
cipher
cryptography
openai
langchain-community
langchain-core
langchain
langchain-openai
faiss-cpu
sentence-transformers